Introduce Gransy provider.
Gransy provider is former Subreg provider, but because it works for
multiple sites (subreg.cz, regtons.com and regnames.eu), Gransy as
a company name is selected as a provider name.

Subreg provider is now a simple proxy that uses Gransy provider for
backwards compatibility.
