+++ README.md
@@ -54,6 +54,7 @@ The current supported providers are:
 - Glesys ([docs](https://github.com/glesys/API/wiki/))
 - GoDaddy ([docs](https://developer.godaddy.com/getstarted#access))
 - Google Cloud DNS ([docs](https://cloud.google.com/dns/api/v1/))
+- Gransy (sites subreg.cz, regtons.com and regnames.eu, [docs](https://subreg.cz/manual/))
 - Hover ([docs](https://hoverapi.docs.apiary.io/))
 - Hurricane Electric DNS ([docs](https://dns.he.net/))
 - Hetzner ([docs](https://wiki.hetzner.de/index.php/DNS_Zonendatei/en))
@@ -81,7 +82,7 @@ The current supported providers are:
 - Sakura Cloud by SAKURA Internet Inc. ([docs](https://developer.sakura.ad.jp/cloud/api/1.1/))
 - SafeDNS by UKFast ([docs](https://developers.ukfast.io/documentation/safedns))
 - SoftLayer ([docs](https://sldn.softlayer.com/article/REST#HTTP_Request_Types))
-- Subreg ([docs](https://subreg.cz/manual/))
+- Subreg (deprecated, use Gransy)
 - Transip ([docs](https://www.transip.nl/transip/api/))
 - Vultr ([docs](https://www.vultr.com/api/))
 - Yandex ([docs](https://tech.yandex.com/domain/doc/reference/dns-add-docpage/))
