+++ lexicon/tests/providers/test_gransy.py
@@ -1,4 +1,4 @@
-"""Integration tests for Subreg"""
+"""Integration tests for Gransy"""
 from unittest import TestCase
 
 from lexicon.tests.providers.integration_tests import IntegrationTests
@@ -7,7 +7,7 @@ from lexicon.tests.providers.integration_tests import IntegrationTests
 # Hook into testing framework by inheriting unittest.TestCase and reuse
 # the tests which *each and every* implementation of the interface must
 # pass, by inheritance from integration_tests.IntegrationTests
-class SubregProviderTests(TestCase, IntegrationTests):
-    """TestCase for Subreg"""
-    provider_name = 'subreg'
+class GransyProviderTests(TestCase, IntegrationTests):
+    """TestCase for Gransy"""
+    provider_name = 'gransy'
     domain = 'oldium.xyz'
