+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_authenticate.yaml
