+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_authenticate_with_unmanaged_domain_should_fail.yaml
