+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_when_calling_create_record_for_CNAME_with_valid_name_and_content.yaml
