+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_when_calling_create_record_for_TXT_with_full_name_and_content.yaml
