+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_when_calling_delete_record_by_filter_should_remove_record.yaml
