+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_when_calling_delete_record_with_record_set_name_remove_all.yaml
