+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_when_calling_list_records_with_fqdn_name_filter_should_return_record.yaml
