+++ tests/fixtures/cassettes/gransy/IntegrationTests/test_provider_when_calling_list_records_with_no_arguments_should_list_all.yaml
